requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
idna==3.13
pyserial==3.5
requests==2.32.5
urllib3==2.6.3
